Fix and split also test for fmod
